requirements.txt
instabot==0.117.0
python-dotenv==0.12.0